refactor: consolidate viem support into EIP4361AuthProvider and add the related type guards to the shared package
