+++ packages/taco/src/encrypt-decrypt.ts
@@ -1,5 +1,12 @@
 import { DkgPublicKey, ThresholdMessageKit } from '@nucypher/nucypher-core';
-import { type Account, Domain, type PublicClient } from '@nucypher/shared';
+import {
+  Account,
+  type Domain,
+  isViemAccount,
+  isViemClient,
+  PublicClient,
+  validateProviderSignerCompatibility,
+} from '@nucypher/shared';
 import { ethers } from 'ethers';
 
 import { Condition } from './conditions/condition';
@@ -47,7 +54,7 @@ export async function encrypt(
   validateProviderSignerCompatibility(providerOrClient, signerOrAccount);
 
   // Type guard to determine if we're using viem or ethers
-  if (isViemClient(providerOrClient)) {
+  if (isViemClient(providerOrClient) && isViemAccount(signerOrAccount)) {
     return viemEncrypt(
       providerOrClient as PublicClient,
       domain,
@@ -149,60 +156,6 @@ export async function decrypt(
       messageKit,
       context,
       porterUris,
-    );
-  }
-}
-
-// Type guard to determine if the client is a viem PublicClient
-function isViemClient(
-  client: ethers.providers.Provider | PublicClient,
-): client is PublicClient {
-  const hasViemProperties = 'chain' in client;
-  const hasViemMethods =
-    typeof (client as { getChainId: () => Promise<number> }).getChainId ===
-    'function';
-  const isNotEthersProvider = !(
-    client instanceof ethers.providers.BaseProvider
-  );
-
-  return isNotEthersProvider && (hasViemProperties || hasViemMethods);
-}
-
-// Type guard to determine if the signer is a viem Account
-function isViemAccount(signer: ethers.Signer | Account): signer is Account {
-  // Check for viem Account properties
-  const hasViemAccountProperties =
-    'address' in signer &&
-    typeof (signer as { address: string }).address === 'string' &&
-    !('provider' in signer); // ethers.Signer has provider property
-
-  // Check if it's not an ethers.Signer
-  const isNotEthersSigner = !(signer instanceof ethers.Signer);
-
-  return isNotEthersSigner && hasViemAccountProperties;
-}
-
-// Validate that provider and signer types are compatible
-function validateProviderSignerCompatibility(
-  providerOrClient: ethers.providers.Provider | PublicClient,
-  signerOrAccount: ethers.Signer | Account,
-): void {
-  const isViemProvider = isViemClient(providerOrClient);
-  const isViemSigner = isViemAccount(signerOrAccount);
-
-  if (isViemProvider && !isViemSigner) {
-    throw new Error(
-      'Type mismatch: viem PublicClient provided but ethers.Signer detected. ' +
-        'When using viem, please provide a viem Account. ' +
-        'Use either: (ethers.Provider + ethers.Signer) or (viem.PublicClient + viem.Account)',
-    );
-  }
-
-  if (!isViemProvider && isViemSigner) {
-    throw new Error(
-      'Type mismatch: ethers.Provider provided but viem Account detected. ' +
-        'When using ethers, please provide an ethers.Signer. ' +
-        'Use either: (ethers.Provider + ethers.Signer) or (viem.PublicClient + viem.Account)',
     );
   }
 }
